backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
python-jose
pydantic
scikit-learn
pandas
numpy
joblib
python-dotenv
bcrypt==3.2.0